+++ packages/go/luna/telemetry/logger.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"os"
 	"regexp"
-	"strings"
+
 	"time"
 )